+++ app/globals.css
@@ -49,20 +49,15 @@ body {
 a {
   color: inherit;
   text-decoration: none;
+  -webkit-tap-highlight-color: transparent;
+}
+
+button {
+  -webkit-tap-highlight-color: transparent;
 }
 
 :root:has(.disable-body-scroll) {
   overflow: hidden;
-}
-
-#contact-cards-container {
-  scroll-margin-top: 120px;
-}
-
-@media (max-width: 900px) {
-  #contact-cards-container {
-    scroll-margin-top: 20px;
-  }
 }
 
 /* Banner Carousel styles */
