+++ app/products/women/page.tsx
@@ -1,13 +1,19 @@
 import { Suspense } from "react";
 import { getWomensProducts } from "@/app/utils/getWomensProducts";
-import HomeClient from "./HomeClient";
+import WomenProductsPage from "./WomenProductsClient";
+import { Metadata } from "next";
 
-export default async function HomePage() {
+export const metadata: Metadata = {
+    title: "Women's Collection - Sewing Patterns | sewmii",
+    description: "Browse our premium women's sewing patterns for corsets, dresses, tops, and more. Perfect for home sewists and designers."
+};
+
+export default async function Page() {
     const { corsets } = await getWomensProducts();
 
     return (
         <Suspense fallback={<div>Loading...</div>}>
-            <HomeClient initialCorsets={corsets} />
+            <WomenProductsPage initialCorsets={corsets} />
         </Suspense>
     );
 }
